Added Security Fixes

+++ web/app.py
@@ -1,5 +1,5 @@
 from base64 import b64encode
-from flask import Flask, render_template, request, Blueprint, redirect, url_for, flash
+from flask import Flask, render_template, request, Blueprint, redirect, url_for, flash, session
 from flask_bootstrap import Bootstrap
 from flask_nav import Nav
 from flask_nav.elements import *
@@ -11,6 +11,7 @@ from flask_images import Images
 from wtforms import TextField, HiddenField, ValidationError, RadioField,\
     BooleanField, SubmitField, PasswordField, FormField, validators
 import psycopg2, os
+import hashlib
 
 
 nav = Nav()
@@ -19,13 +20,12 @@ app.secret_key = 'howareyoudoingthisisasecretkeyformyapp'
 Material(app)
 Images(app)
 nav.init_app(app)
-app.config['WTF_CSRF_ENABLED'] = False
 app.config['IMAGES_URL'] = ''
 
 
 #return a connection and cursor for db
 def get_database():
-    conn = psycopg2.connect("dbname='flask' user='flask' host='db'")
+    conn = psycopg2.connect("dbname='flask' user='flask' host='db' password='flaskdb'")
     return conn, conn.cursor()
 #validator to make sure username doesn't exist
 def val_username():
@@ -36,6 +36,14 @@ def val_username():
             con.close()
             raise ValidationError("Username is already taken")
     return _val_username
+
+def val_image(message=u'Images only!', extensions=None):
+    if not extensions:
+        extensions = ('jpg', 'jpeg', 'png', 'gif')
+    def _val_image(form, field):
+        if not field.data or field.data.filename.split('.')[-1].lower() not in extensions:
+            raise ValidationError(message)
+    return _val_image
 
 #Login Form Class with custom validation to check for username/password match:
 class LoginForm(FlaskForm): 
@@ -50,7 +58,7 @@ class LoginForm(FlaskForm):
         cursor.execute('SELECT password FROM public.users WHERE username=%s', (self.username.data,))
         check = cursor.fetchone()
         con.close()
-        if check == None or (check != None and check[0] != self.password.data):
+        if check == None or (check != None and check[0] != hashlib.sha512(self.password.data.encode()).hexdigest()):
             self.password.errors.append("Incorrect Login")
             return False
         return True
@@ -65,14 +73,17 @@ class RegisterForm(FlaskForm):
         rv = FlaskForm.validate(self)
         if not rv:
             return False
-
-        
         return True
 #Photo Upload Form
 class PhotoForm(FlaskForm):
-    photo = FileField('Sample upload', [validators.required()])
+    photo = FileField('Sample upload', [val_image(), validators.required()])
     caption = TextField('Caption')
     submit_button = SubmitField('Submit Form')
+    def validate(self):
+        rv = FlaskForm.validate(self)
+        if not rv:
+            return False
+        return True
 
 #Form for editing photo captions
 class EditForm(FlaskForm):
@@ -112,22 +123,16 @@ def login():
     form = LoginForm()
     if request.method == 'POST' and form.validate_on_submit():
         username = request.values.get('username')
-        password = request.values.get('password')
-        con, cur = get_database()
-        cur.execute('SELECT password FROM public.users WHERE username=%s', (username,)) #fetches password from database for a user
-        check = cur.fetchone()
-        con.close()
-        if check != None and check[0] == password:
-            response = redirect(url_for('home'))
-            response.set_cookie('USERNAME', username) #sets username cookie
-            return response
+        response = redirect(url_for('home'))
+        session['username'] = username #sets username cookie
+        return response
     return render_template('login.html', form=form)
 
 #Removes cookie effectively logging a user out
 @app.route('/logout', methods=['GET'])
 def logout():
     response = redirect(url_for('home'))
-    response.set_cookie('USERNAME','', expires=0)
+    session.pop('username', None)
     return response
 
 #Path to allow a user to register for an account, provided that the username doesn't already exist
@@ -136,8 +141,9 @@ def register():
     form = RegisterForm()
     if request.method == 'POST' and form.validate_on_submit():
         con, cursor = get_database()
+        password = hashlib.sha512(request.values.get('password').encode()).hexdigest()
         cursor.execute("""INSERT INTO public.users VALUES(%s, %s)""", (request.values.get('username')
-            , request.values.get('password'))) #inserts new username and password into the db
+            , password)) #inserts new username and password into the db
         con.commit()
         con.close()
         return redirect(url_for('login'))
@@ -146,8 +152,9 @@ def register():
 #Only available to users, allows them to upload new images using the PhotoForm
 @app.route('/upload', methods=['GET','POST'])
 def upload():
-    username = request.cookies.get('USERNAME')
-    if request.method == 'POST' and username:
+    form = PhotoForm()
+    if request.method == 'POST' and 'username' in session and form.validate_on_submit():
+        username = session['username']
         caption = request.values.get('caption')
         con, cursor = get_database()
         image = request.files.get('photo')
@@ -158,14 +165,13 @@ def upload():
         new.close()
         con.commit()
         con.close()
-    form = PhotoForm()
     return render_template('upload.html', form=form)
 
 #Allows users to edit their captions for an image
 @app.route('/edit', methods=['GET', 'POST'])
 def edit():
     oid = request.values.get('oid')
-    if request.method == 'POST':
+    if request.method == 'POST' and 'username' in session:
         new_caption = request.values.get('caption')
         con, cursor = get_database()
         cursor.execute("""UPDATE public.images SET caption=%s WHERE oid=%s""", (new_caption, oid)) #updates image table with new caption
@@ -186,7 +192,7 @@ def edit():
 @app.route('/delete', methods=['GET'])
 def delete():
     oid = request.values.get('oid')
-    if oid:
+    if oid and 'username' in session:
         con, cursor = get_database()
         cursor.execute("""DELETE FROM public.images WHERE oid=%s""", (oid,)) #deleted image oid from db, markes image ready for deletion
         con.commit()
@@ -221,7 +227,9 @@ def home(page=1):
         else:
             images.append((Image(path1, caption1),)) #if total number of images is odd, then the last tuple only has one Image Class object
     con.close()
-    username = request.cookies.get('USERNAME') #to see if its a logged in user or a visitor
+    username = None
+    if 'username' in session:
+        username = session['username'] #to see if its a logged in user or a visitor
     pagination = Pagination(page=page, per_page=10, total=int(result[0]), search=False, record_name='images')
     return render_template('home.html', images=images, per_page=10, pagination=pagination, username=username)
 
@@ -229,7 +237,7 @@ def home(page=1):
 #Most of the code is the same as the home function save for the query which also adds the username constraint for fetched items
 @app.route('/myimages')
 def myimages(page=1):
-    username = request.cookies.get('USERNAME')
+    username = session['username']
     page, per_page, offset = get_page_args(page_parameter='page',
                                            per_page_parameter='per_page')
     con, cursor = get_database()
